Move librustc panic handler into the new one
